Bugfix: There was a problem with the formular for editing users which have been fixed in version 2.2.1
In all versions before 2.2 the template parameter $FORM['DATA']['PASSWORD'] was explicitly given, but always set to the value "NULL". Due to the changes in version 2.2, the parameter is no longer passed at all, which caused PHP to issue an notice via an undefined array index. The problem has now been solved by removing the "value" attribute for the password field from the form.