Fix javascript problem that causes unintended behavior
If you browse in a browser tab through some websites and want to go back to the previous page, post or user, you can either click the history back button, press the corresponding key on your mouse or even press <alt>+<arrow-left> on your keyboard. But if you only press <arrow-left> on a page, post or user, the callback function of the "keyup" event will change the location in the desired direction on the blog if there is a previous page, post or user, and if neither <ctrl> nor <shift> is pressed. This is a functionality of the javascript from the template and not a functionality of your browser.

Pressing <alt>+<arrow-left> IS a functionality of the browser and should not conflict with the behavior when <arrow-left> is pressed alone. This commit fixes this problem by adding the condition that <alt> should not be pressed either.