Use grid layout for items in administration area
Introduce a new 1/2/3-column CSS grid layout in the administration area
for items on overview pages and in the search results. The column count
of the grid depends on the users viewport width.

In addition, the default value of the following configuration settings
has been changed to 12 because 12 can also be divided by 2 and 3 which
is useful for the 1/2/3-column grid layout.

ADMIN.PAGE.LIST_SIZE = 12
ADMIN.POST.LIST_SIZE = 12