Remove migration for migration table (readme)
This commit removes the migration number 5 which created the migration
table. The initialization of the migration table will be implemented in
the upcoming migration script by itself. Sorry!

If you recently installed the application with a fresh database which
contained the migration table, please manually set the schema_version
to 4. It is recommended to only install a specific tag instead of the
latest master branch unless you are a developer or know what you do.

UPDATE migration SET schema_version = 4