Add option CATEGORY.LIST_SORT to application.php
The option CATEGORY.LIST_SORT is currently only used for sorting the
children categories of a category, not for sorting the categories on
the category overview page.

This is because the sorting of the categories in the category overview
is handelnd internally with a virtual table column "_depth" which sorts
the categories by name *and* its sub categories (like a tree).

Thus, there is currently no user option to sort the category overview.