Add dark color mode for admin theme
This commit adds a dark color mode for the admin theme. The dark color
mode can be enabled/disabled by clicking the hyperlink in the footer.

It is not perfect yet (it uses cookies and needs a full page reload),
but it works just fine for the moment. Hope your eyes can enjoy it!