Replace option $ITEM.EMOTICONS with WRAP_EMOTICONS
This commit adds the configuration option "WRAP_EMOTICONS" to supersede
the configuration options "(CATEGORY|PAGE|POST|USER).EMOTICONS".

If the value of "WRAP_EMOTICONS" evaluates to TRUE, the emoticons found
in the content of an item will be wrapped into a "<span>" element which
gives the user a description of the emoticon in the "title" attribute.