Create feed item's GUID manually in template file
Do not rely on the "$POST['GUID']" template parameter anymore and create
the GUID for the feed item manually in the feed item's template file.

The application internal function "generatePseudoGUID" has been removed
and the function "generateItemTemplateData" will not return the "GUID"
part anymore (which was only used by the feed item template anyway).

For backward compatibility, the "$POST['GUID']" template parameter will
still be present in the feed item template, but not in other templates!