Hotfix: Replace `"` with `'` in link title string
Because there currently is no sane way to escape double quotes within a
string intended to be used as title for a Markdown formatted link, just
replace the double with single quotes until a better solution is found.

Note: Just replacing with `&quot;` will not work here because Parsedown
escapes this further to `&amp;quot;`.