FunctionParser: No static property/methods anymore
Problem: When creating a new "FunctionParser" object, it should not know
the previously registered content functions which were saved in a static
class property and registered via static class method.

If one creates a new "FunctionParser" object, one would expect that the
array of registered content functions in that object is empty.

Changes:
- Make static property "functions" non-static
- Make static method "register" non-static
- Add new method "registerFromArray"