Rename "rsrc" directories to "static"
Stop using "rsrc" as name for the directories which contain static files
and rename them to "static". This affects the global static directory
and the subdirectories in the theme directories.

If you've stored your own files in the old global "rsrc" directory, you
must move them to the new "static" directory.